Deleted src/tests and moved mock data into the features directory.
