+++ src/features/common/mocks/authorizationMock.ts
@@ -1,0 +1,7 @@
+// tests/mocks/authorizationMock.ts
+export const mockVerifyUser = jest.fn();
+
+jest.mock("@features/common/authorization.service", () => ({
+	...jest.requireActual("@/features/common/authorization.service"),
+	verifyUser: mockVerifyUser,
+}));
