+++ src/features/common/mocks/datas/store.ts
@@ -1,0 +1,16 @@
+import { faker } from "@faker-js/faker";
+import type { Store } from "@prisma/client";
+
+export const createMockStoreInput = () => ({
+	name: faker.company.name(),
+	groupId: faker.string.uuid(),
+});
+
+export const mockStore: Store = {
+	id: faker.string.uuid(),
+	storeId: faker.string.uuid(),
+	name: faker.company.name(),
+	groupId: faker.string.nanoid(),
+	createdAt: new Date(),
+	updatedAt: new Date(),
+};
